+++ web_app/order_books/ekubo/main.py
@@ -1,57 +1,15 @@
 from decimal import Decimal, getcontext
 import pandas as pd
-from dataclasses import dataclass
-from web_app.ekubo.api_connector import EkuboAPIConnector
+from order_books.abstractions import OrderBookBase
+from web_app.order_books.ekubo.api_connector import EkuboAPIConnector
 from functools import partial
 
 
 getcontext().prec = 18
 
 
-@dataclass
-class TokenConfig:
-    name: str
-    decimals: int
-
-
-TOKEN_MAPPING = {
-    "0x49d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7": TokenConfig(
-        name="ETH", decimals=18
-    ),
-    "0x53c91253bc9682c04929ca02ed00b3e423f6710d2ee7e0d5ebb06f3ecf368a8": TokenConfig(
-        name="USDC", decimals=6
-    ),
-    "0x05fa6cc6185eab4b0264a4134e2d4e74be11205351c7c91196cb27d5d97f8d21": TokenConfig(
-        name="USDT", decimals=6
-    ),
-    "0x019c981ec23aa9cbac1cc1eb7f92cf09ea2816db9cbd932e251c86a2e8fb725f": TokenConfig(
-        name="DAI", decimals=18
-    ),
-    "0x01320a9910e78afc18be65e4080b51ecc0ee5c0a8b6cc7ef4e685e02b50e57ef": TokenConfig(
-        name="wBTC", decimals=8
-    ),
-    "0x07514ee6fa12f300ce293c60d60ecce0704314defdb137301dae78a7e5abbdd7": TokenConfig(
-        name="STRK", decimals=18
-    ),
-    "0x0719b5092403233201aa822ce928bd4b551d0cdb071a724edd7dc5e5f57b7f34": TokenConfig(
-        name="UNO", decimals=18
-    ),
-    "0x00585c32b625999e6e5e78645ff8df7a9001cf5cf3eb6b80ccdd16cb64bd3a34": TokenConfig(
-        name="ZEND", decimals=18
-    ),
-    "0x07e2c010c0b381f347926d5a203da0335ef17aefee75a89292ef2b0f94924864": TokenConfig(
-        name="wstETH", decimals=18
-    ),
-    "0x4c4fb1ab068f6039d5780c68dd0fa2f8742cceb3426d19667778ca7f3518a9": TokenConfig(
-        name="LORDS", decimals=18
-    ),
-}
-
-
-class EkuboOrderBook:
+class EkuboOrderBook(OrderBookBase):
     DEX = "Ekubo"
-    MIN_PRICE_RANGE = Decimal("0.1")
-    MAX_PRICE_RANGE = Decimal("1.90")
 
     def __init__(self, token_a: str, token_b: str):
         """
@@ -59,16 +17,7 @@ class EkuboOrderBook:
         :param token_a: BaseToken contract address
         :param token_b: QuoteToken contract address
         """
-        self.token_a = token_a
-        self.token_b = token_b
-        self.asks = []  # List of tuples (price, quantity)
-        self.bids = []  # List of tuples (price, quantity)
-        self.timestamp = None
-        self.block = None
-        self.tick_current_price = Decimal("0")
-        self.token_a_decimal = TOKEN_MAPPING.get(token_a).decimals
-        self.token_b_decimal = TOKEN_MAPPING.get(token_b).decimals
-        self.total_liquidity = Decimal("0")
+        super().__init__(token_a, token_b)
         self.connector = EkuboAPIConnector()
 
     @property
@@ -181,15 +130,6 @@ class EkuboOrderBook:
                 bid_data.append(sorted_data)
         return ask_data, bid_data
 
-    def calculate_price_range(self) -> tuple:
-        """
-        Calculate the minimum and maximum price based on the current price.
-        :return: tuple - The minimum and maximum price range.
-        """
-        min_price = self.tick_current_price * self.MIN_PRICE_RANGE
-        max_price = self.tick_current_price * self.MAX_PRICE_RANGE
-        return min_price, max_price
-
     def calculate_liquidity_amount(
         self, tick: Decimal, liquidity_pair_total: Decimal
     ) -> Decimal:
@@ -203,14 +143,6 @@ class EkuboOrderBook:
         liquidity_delta = liquidity_pair_total / (sqrt_ratio / Decimal(2**128))
         return liquidity_delta / 10**self.token_a_decimal
 
-    def get_sqrt_ratio(self, tick: Decimal) -> Decimal:
-        """
-        Get the square root ratio based on the tick.
-        :param tick: tick value
-        :return: square root ratio
-        """
-        return (Decimal("1.000001").sqrt() ** tick) * (Decimal(2) ** 128)
-
     def tick_to_price(self, tick: Decimal) -> Decimal:
         """
         Convert tick to price.
@@ -223,17 +155,6 @@ class EkuboOrderBook:
             self.token_a_decimal - self.token_b_decimal
         )
         return price
-
-    def get_order_book(self) -> dict:
-        return {
-            "token_a": self.token_a,
-            "token_b": self.token_b,
-            "timestamp": self.timestamp,
-            "block": self.block,
-            "dex": self.DEX,
-            "asks": sorted(self.asks, key=lambda x: x[0]),
-            "bids": sorted(self.bids, key=lambda x: x[0]),
-        }
 
     def calculate_price_change(self, sell_amount: Decimal) -> tuple:
         remaining_sell_amount = sell_amount
